inference/requirements.txt
torch
torch_geometric
numpy
scikit-learn
google-generativeai
python-dotenv
python-hcl2
